+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+## Added
+
+- Added `deactivate` command.
+
 ## [0.1.2] - 2020-08-01Z
 
 ## Fixed